# GG-9341

+++ modules/core/src/test/java/org/gridgain/grid/session/GridSessionFutureWaitJobAttributeSelfTest.java
@@ -235,7 +235,7 @@ public class GridSessionFutureWaitJobAttributeSelfTest extends GridCommonAbstrac
 
             for (GridComputeJobResult result : results) {
                 if (result.getData() != null)
-                    sum += result.getData();
+                    sum += (Integer)result.getData();
             }
 
             return sum;
